feat(Complex karma config\): allow complex karma config
Remove the `KarmaConfigEditor`. It is not needed anymore
since there is no need to edit Stryker's `files` array. This allows
us to read karma configuration from the test runner child
process instead of the main process.
