+++ packages/stryker-karma-runner/src/index.ts
@@ -1,7 +1,4 @@
 import { TestRunnerFactory } from 'stryker-api/test_runner';
-import { ConfigEditorFactory } from 'stryker-api/config';
 import KarmaTestRunner from './KarmaTestRunner';
-import KarmaConfigEditor from './KarmaConfigEditor';
 
 TestRunnerFactory.instance().register('karma', KarmaTestRunner);
-ConfigEditorFactory.instance().register('karma', KarmaConfigEditor);
